support check specified files is change to update

+++ utils/index.js
@@ -7,33 +7,98 @@ export const createWorker = (func) => {
 };
 export const createWorkerFunc = () => {
     let oldVersion = '';
+    const flagSet = new Set();
     let intervalTime = 5000;
     let immediate = false;
     let originFileUrl = '';
+    let checkOriginSpecifiedFilesUrl = [];
+    let checkOriginSpecifiedFilesUrlMode = 'one';
+    // TODO: 类型定义
+    let timer = null;
+    let clearIntervalOnDialog = false;
     const temp = self;
     temp.onmessage = (event) => {
         oldVersion = event.data['version-key'];
         intervalTime = event.data['polling-time'];
         immediate = event.data['immediate'];
         originFileUrl = event.data['origin-version-file-url'];
+        checkOriginSpecifiedFilesUrl =
+            event.data['check-origin-specified-files-url'];
+        checkOriginSpecifiedFilesUrlMode =
+            event.data['check-origin-specified-files-url-mode'];
+        clearIntervalOnDialog = event.data['clear-interval-on-dialog'];
+        const checkVersionType = oldVersion && originFileUrl
+            ? 'check-version'
+            : (checkOriginSpecifiedFilesUrl === null || checkOriginSpecifiedFilesUrl === void 0 ? void 0 : checkOriginSpecifiedFilesUrl.length)
+                ? 'check-specified-files'
+                : '';
+        if (!checkVersionType)
+            return console.warn('Not found localPackageVersion, originVersionFileUrl or originSpecifiedFilesUrl');
+        console.log('You are use check version type is', checkVersionType);
         const doFetch = () => {
-            fetch(`${originFileUrl}?${+new Date()}`)
-                .then((res) => {
-                return res.json();
-            })
-                .then((versionJsonFile) => {
-                if (oldVersion !== versionJsonFile.version) {
-                    temp.postMessage({
-                        refreshPageVersion: `${versionJsonFile.version}`,
-                        external: versionJsonFile.external,
+            if (checkVersionType === 'check-version') {
+                fetch(`${originFileUrl}?${+new Date()}`)
+                    .then((res) => {
+                    return res.json();
+                })
+                    .then((versionJsonFile) => {
+                    if (oldVersion !== versionJsonFile.version) {
+                        temp.postMessage({
+                            refreshPageVisible: true,
+                            refreshPageVersion: `${versionJsonFile.version}`,
+                            external: versionJsonFile.external,
+                        });
+                    }
+                });
+            }
+            if (checkVersionType === 'check-specified-files') {
+                if (!(checkOriginSpecifiedFilesUrl === null || checkOriginSpecifiedFilesUrl === void 0 ? void 0 : checkOriginSpecifiedFilesUrl.length))
+                    return;
+                checkOriginSpecifiedFilesUrl.forEach((url) => {
+                    fetch(url)
+                        .then((res) => {
+                        return res.headers.get('ETag') || res.headers.get('Last-Modified');
+                    })
+                        .then((flag) => {
+                        if (flag) {
+                            console.log('flag', flag, flagSet);
+                            flagSet.add(flag);
+                        }
+                        // TODO: 第一次轮训不生效, 需要第二次轮训才生效, 需要考虑如果用户的循环时间很长怎么办
+                        if (checkOriginSpecifiedFilesUrlMode === 'one' &&
+                            (flagSet === null || flagSet === void 0 ? void 0 : flagSet.size) > (checkOriginSpecifiedFilesUrl === null || checkOriginSpecifiedFilesUrl === void 0 ? void 0 : checkOriginSpecifiedFilesUrl.length)) {
+                            temp.postMessage({
+                                refreshPageVisible: true,
+                                refreshPageVersion: '',
+                                // TODO: 本版本的修改内容
+                                external: '',
+                            });
+                            flagSet.clear();
+                            if (clearIntervalOnDialog) {
+                                clearInterval(timer);
+                            }
+                        }
+                        if (checkOriginSpecifiedFilesUrlMode === 'all' &&
+                            (flagSet === null || flagSet === void 0 ? void 0 : flagSet.size) === (checkOriginSpecifiedFilesUrl === null || checkOriginSpecifiedFilesUrl === void 0 ? void 0 : checkOriginSpecifiedFilesUrl.length) * 2) {
+                            temp.postMessage({
+                                refreshPageVisible: true,
+                                refreshPageVersion: '',
+                                // TODO: 本版本的修改内容
+                                external: '',
+                            });
+                            flagSet.clear();
+                            if (clearIntervalOnDialog) {
+                                clearInterval(timer);
+                            }
+                        }
                     });
-                }
-            });
+                });
+            }
         };
         if (immediate) {
             doFetch();
         }
-        setInterval(doFetch, intervalTime);
+        timer = setInterval(doFetch, intervalTime);
     };
     return temp;
 };
@@ -46,11 +111,10 @@ export const cancelUpdateFunc = (cancelMode, newVersion, cancelUpdateAndStopWork
     const isStopWorker = cancelUpdateAndStopWorker || false;
     switch (cancelModeType) {
         case 'ignore-current-version':
-            if (cancelModeTypeValue === newVersion) {
+            if (newVersion && cancelModeTypeValue === newVersion) {
                 isStopWorker && (worker === null || worker === void 0 ? void 0 : worker.terminate());
                 return true;
             }
-            break;
         case 'ignore-today':
             if (cancelModeTypeValue === todayDate) {
                 isStopWorker && (worker === null || worker === void 0 ? void 0 : worker.terminate());
